@@ -1,11 +1,6 @@
 module.exports = {
-  plugins: [
-    require("tailwindcss"),
-    require("autoprefixer"),
-    process.env.NODE_ENV === "production" &&
-      require("@fullhuman/postcss-purgecss")({
-        content: ["./src/**/*.tsx", "./src/**/*.ts", "./public/index.html"],
-        defaultExtractor: (content) => content.match(/[A-Za-z0-9-_:/]+/g) || []
-      })
-  ]
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 };
