@@ -1,7 +1,7 @@
-import { addField } from "react-fluent-form";
+import { addField } from 'react-fluent-form'
 
-import { Nullable } from "../../utils/Nullable";
-import { ReactSingleSelect } from "./ReactSingleSelect";
+import { Nullable } from '../../utils/Nullable'
+import { ReactSingleSelect } from './ReactSingleSelect'
 
 declare module "react-fluent-form" {
   interface FieldCreator {
