@@ -1,27 +1,26 @@
-import { motion } from "framer-motion";
-import React, { useContext } from "react";
-import { useFluentForm } from "react-fluent-form";
+import { motion } from 'framer-motion'
+import { useContext } from 'react'
+import { useFluentForm } from 'react-fluent-form'
 
-import { resetPasswordFormConfig } from "../../form-config/ResetPasswordFormConfig";
-import { AuthServiceContext } from "../../service/auth/AuthServiceContext";
-import { createErrorToast } from "../../utils/toast/errorToast";
-import { scaleVariation } from "../common/variants/framerVariants";
+import { resetPasswordFormConfig } from '../../form-config/ResetPasswordFormConfig'
+import { AuthServiceContext } from '../../service/auth/AuthServiceContext'
+import { createErrorToast } from '../../utils/toast/errorToast'
+import { scaleVariation } from '../common/variants/framerVariants'
 
 interface ResetPasswordModalContentProps {
   closeFunction: () => void;
   token: string;
 }
 
-export const ResetPasswordModalContent: React.FC<ResetPasswordModalContentProps> = ({
-  closeFunction,
-  token
-}) => {
+export const ResetPasswordModalContent: React.FC<
+  ResetPasswordModalContentProps
+> = ({ closeFunction, token }) => {
   const {
     values,
     fields,
     handleSubmit,
     reset: resetForm,
-    errors
+    errors,
   } = useFluentForm(resetPasswordFormConfig);
 
   const { resetPassword, loading } = useContext(AuthServiceContext);
