@@ -1,7 +1,7 @@
-import React from "react";
+import { createContext } from 'react'
 
-import { useAuthService } from "./useAuthService";
+import { useAuthService } from './useAuthService'
 
-export const AuthServiceContext = React.createContext<
+export const AuthServiceContext = createContext<
   ReturnType<typeof useAuthService>
 >(undefined as any);
