@@ -1,15 +1,15 @@
-import axios from "axios";
+import axios from 'axios'
 
-import { apiBaseUrl } from "../../constants/endpoints";
-import { handleErrorResponse } from "./handleErrorResponse";
-import { handleRequest } from "./handleRequest";
+import { apiBaseUrl } from '../../constants/endpoints'
+import { handleErrorResponse } from './handleErrorResponse'
+import { handleRequest } from './handleRequest'
 
 export const instance = axios.create({
   baseURL: apiBaseUrl,
   timeout: 30000,
   headers: {
-    "Content-Type": "application/json"
-  }
+    "Content-Type": "application/json",
+  },
 });
 
 instance.interceptors.response.use((response) => response, handleErrorResponse);
