@@ -1,6 +1,6 @@
-import { AxiosRequestConfig } from "axios";
+import { AxiosRequestConfig } from 'axios'
 
-import { TypedStorage } from "../../utils/typedStorage";
+import { TypedStorage } from '../../utils/typedStorage'
 
 // methods will be called before each request
 export const handleRequest = (config: AxiosRequestConfig) => {
@@ -11,7 +11,7 @@ export const handleRequest = (config: AxiosRequestConfig) => {
 // get token from LS
 const setAuthorizationHeader = (config: AxiosRequestConfig) => {
   if (TypedStorage.accessToken) {
-    config.headers.Authorization = `Bearer ${TypedStorage.accessToken}`;
+    config.headers!.Authorization = `Bearer ${TypedStorage.accessToken}`;
   } else {
     console.warn(
       "Unable to read token from localstorage. Authorization header will not be set."
